+++ lib/models/signal.g.dart
@@ -9,8 +9,10 @@ part of 'signal.dart';
 _$_Signal _$$_SignalFromJson(Map<String, dynamic> json) => _$_Signal(
       id: json['id'] as String,
       name: json['name'] as String,
+      hidden: json['hidden'] as bool? ?? false,
     );
 
 Map<String, dynamic> _$$_SignalToJson(_$_Signal instance) => <String, dynamic>{
       'name': instance.name,
+      'hidden': instance.hidden,
     };