+++ lib/router/routes.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:periodik/router/redirect.dart';
 import 'package:periodik/screens/login_screen.dart';
+import 'package:periodik/screens/settings/settings_screen.dart';
 import 'package:periodik/screens/sign_up_screen.dart';
 import 'package:periodik/screens/signal/signal_view.dart';
 import 'package:periodik/screens/signals_calendar_screen.dart';
@@ -127,5 +128,15 @@ class SignalListRoute extends GoRouteData {
       id: id,
       view: SignalView.list,
     );
+  }
+}
+
+@TypedGoRoute<SettingsRoute>(path: '/settings')
+class SettingsRoute extends GoRouteData {
+  const SettingsRoute();
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const SettingsScreen();
   }
 }