+++ lib/router/routes.g.dart
@@ -11,6 +11,7 @@ List<RouteBase> get $appRoutes => [
       $verifyEmailRoute,
       $loginRoute,
       $signalsRoute,
+      $settingsRoute,
     ];
 
 RouteBase get $homeRoute => GoRouteData.$route(
@@ -207,6 +208,28 @@ extension $SignalListRouteExtension on SignalListRoute {
 
   String get location => GoRouteData.$location(
         '/signals/${Uri.encodeComponent(id)}/list',
+      );
+
+  void go(BuildContext context) => context.go(location);
+
+  Future<T?> push<T>(BuildContext context) => context.push<T>(location);
+
+  void pushReplacement(BuildContext context) =>
+      context.pushReplacement(location);
+
+  void replace(BuildContext context) => context.replace(location);
+}
+
+RouteBase get $settingsRoute => GoRouteData.$route(
+      path: '/settings',
+      factory: $SettingsRouteExtension._fromState,
+    );
+
+extension $SettingsRouteExtension on SettingsRoute {
+  static SettingsRoute _fromState(GoRouterState state) => const SettingsRoute();
+
+  String get location => GoRouteData.$location(
+        '/settings',
       );
 
   void go(BuildContext context) => context.go(location);