+++ lib/screens/signals_calendar_screen.dart
@@ -12,10 +12,6 @@ import '../models/point_state.dart';
 import '../providers/display_predictions_provider.dart';
 import '../utils/hero_tags.dart';
 
-enum _Action {
-  displayPredictions,
-}
-
 class SignalsCalendarScreen extends ConsumerStatefulWidget {
   const SignalsCalendarScreen({super.key});
 
@@ -27,14 +23,6 @@ class SignalsCalendarScreen extends ConsumerStatefulWidget {
 class __SignalsCalendarScreenState
     extends ConsumerState<SignalsCalendarScreen> {
   final _controller = CalendarController();
-
-  void _onSelected(_Action action) {
-    switch (action) {
-      case _Action.displayPredictions:
-        ref.read(displayPredictionsProvider.notifier).update((state) => !state);
-        break;
-    }
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -62,23 +50,6 @@ class __SignalsCalendarScreenState
     return Scaffold(
       appBar: AppBar(
         title: const Text('Signals Calendar'),
-        actions: [
-          PopupMenuButton<_Action>(
-            onSelected: _onSelected,
-            itemBuilder: (context) => [
-              PopupMenuItem(
-                value: _Action.displayPredictions,
-                child: Consumer(builder: (context, ref, child) {
-                  final displayPredictions =
-                      ref.watch(displayPredictionsProvider);
-                  return Text(
-                    '${displayPredictions ? 'Hide' : 'Display'} predictions',
-                  );
-                }),
-              ),
-            ],
-          )
-        ],
       ),
       body: Center(
         child: Calendar(